added onednn_inductor_quantizer to remove redundant code in arm_inductor_quantizer and x86_inductor_quantizer
