+++ test/quantization/pt2e/test_arminductor_quantizer.py
@@ -12,16 +12,16 @@ from torch.ao.quantization.quantize_pt2e import (
     prepare_pt2e,
     prepare_qat_pt2e,
 )
-from torch.ao.quantization.quantizer.arm_inductor_quantizer import (
+from torch.ao.quantization.quantizer.arm_inductor_quantizer import ArmInductorQuantizer
+from torch.ao.quantization.quantizer.onednn_inductor_quantizer import (
     QUANT_ANNOTATION_KEY,
-    ArmInductorQuantizer,
 )
 from torch.export import export_for_training
 from torch.testing._internal.common_quantization import (
     NodeSpec as ns,
     QuantizationTestCase,
-    skipIfNoInductorSupport,
     skipIfNoArm,
+    skipIfNoInductorSupport,
 )
 from torch.testing._internal.common_quantized import override_quantized_engine
 from torch.testing._internal.common_utils import skipIfTorchDynamo
@@ -45,25 +45,6 @@ class TestHelperModules:
             x = self.conv(x)
             if self.with_bn:
                 x = self.bn(x)
-            return x
-
-    class Conv2dUnaryModule(torch.nn.Module):
-        def __init__(self, post_op, use_bias: bool = False, with_bn=False) -> None:
-            super().__init__()
-            self.conv = nn.Conv2d(
-                3, 6, (2, 2), stride=(1, 1), padding=(1, 1), bias=use_bias
-            )
-            self.post_op = post_op
-            self.bn = torch.nn.BatchNorm2d(6)
-            self.with_bn = with_bn
-            self.maxpool = torch.nn.MaxPool2d((3, 3))
-
-        def forward(self, x):
-            x = self.conv(x)
-            if self.with_bn:
-                x = self.bn(x)
-            x = self.post_op(x)
-            x = self.maxpool(x)
             return x
 
     class Conv2dAddModule(torch.nn.Module):
@@ -125,67 +106,6 @@ class TestHelperModules:
                 else:
                     return self.conv(x) + self.conv2(x)
 
-    class Conv2dAddReLUModule(torch.nn.Module):
-        def __init__(
-            self,
-            inplace_add: bool = False,
-            conv2d_type: NodePosType = NodePosType.left,
-            inplace_relu: bool = False,
-            use_bias: bool = False,
-            with_bn: bool = False,
-        ) -> None:
-            super().__init__()
-            self.conv = torch.nn.Conv2d(
-                in_channels=3,
-                out_channels=3,
-                kernel_size=3,
-                stride=1,
-                padding=1,
-                bias=use_bias,
-            )
-            self.conv2 = torch.nn.Conv2d(
-                in_channels=3,
-                out_channels=3,
-                kernel_size=3,
-                stride=1,
-                padding=1,
-                bias=use_bias,
-            )
-            self.relu = nn.ReLU()
-            self.inplace_add = inplace_add
-            self.conv2d_type = conv2d_type
-            self.relu2 = nn.ReLU(inplace=inplace_relu)
-            self.bn = torch.nn.BatchNorm2d(3)
-            self.with_bn = with_bn
-
-        def forward(self, x):
-            if self.conv2d_type == NodePosType.left:
-                if self.inplace_add:
-                    tmp = self.conv(x)
-                    if self.with_bn:
-                        tmp = self.bn(tmp)
-                    tmp += self.relu(x)
-                    return self.relu2(tmp)
-                else:
-                    tmp = self.conv(x)
-                    if self.with_bn:
-                        tmp = self.bn(tmp)
-                    return self.relu2(tmp + self.relu(x))
-            elif self.conv2d_type == NodePosType.right:
-                if self.inplace_add:
-                    tmp = self.relu(x)
-                    tmp += self.conv(x)
-                    return self.relu2(tmp)
-                else:
-                    return self.relu2(self.relu(x) + self.conv(x))
-            elif self.conv2d_type == NodePosType.both:
-                if self.inplace_add:
-                    tmp = self.conv(x)
-                    tmp += self.conv2(x)
-                    return self.relu2(tmp)
-                else:
-                    return self.relu2(self.conv(x) + self.conv2(x))
-
     class Conv2dSingleOpPowModule(nn.Module):
         def __init__(self, single_op):
             super().__init__()
@@ -196,123 +116,6 @@ class TestHelperModules:
             x = self.conv(x)
             x = self.single_op(x)
             return torch.pow(x, 2)
-
-    class SerialsConv2dAddReLUModule(torch.nn.Module):
-        """Serials of 2 Conv2d -> Add -> ReLU Pattern."""
-
-        def __init__(
-            self,
-        ) -> None:
-            super().__init__()
-            self.conv = torch.nn.Conv2d(
-                in_channels=3,
-                out_channels=3,
-                kernel_size=3,
-                stride=1,
-                padding=1,
-                bias=True,
-            )
-            self.conv2 = torch.nn.Conv2d(
-                in_channels=3,
-                out_channels=3,
-                kernel_size=3,
-                stride=1,
-                padding=1,
-                bias=True,
-            )
-            self.conv3 = torch.nn.Conv2d(
-                in_channels=3,
-                out_channels=3,
-                kernel_size=3,
-                stride=1,
-                padding=1,
-                bias=True,
-            )
-            self.conv4 = torch.nn.Conv2d(
-                in_channels=3,
-                out_channels=3,
-                kernel_size=3,
-                stride=1,
-                padding=1,
-                bias=True,
-            )
-            self.relu = nn.ReLU()
-            self.relu2 = nn.ReLU()
-
-        def forward(self, x):
-            x1 = self.conv(x)
-            res1 = self.relu(self.conv2(x1) + self.conv3(x1))
-            res2 = self.relu2(self.conv4(res1) + res1)
-            return res2
-
-    class Conv2dCatMaxpool2d(torch.nn.Module):
-        def __init__(
-            self,
-        ):
-            super().__init__()
-            self.conv = torch.nn.Conv2d(
-                3, 16, 7, bias=True, stride=2, padding=3, dilation=1
-            )
-            self.conv2 = torch.nn.Conv2d(
-                3, 16, 7, bias=True, stride=2, padding=3, dilation=1
-            )
-            self.relu = torch.nn.ReLU()
-            self.maxpool = torch.nn.MaxPool2d(3, stride=2, padding=1)
-            self.conv3 = torch.nn.Conv2d(
-                32, 32, 7, bias=True, stride=2, padding=3, dilation=1
-            )
-
-        def forward(self, x):
-            temp1 = self.relu(self.conv(x))
-            temp2 = self.conv2(x + 1)
-            temp3 = torch.cat((temp1, temp2), 1)
-            temp4 = self.maxpool(temp3)
-            temp5 = self.conv3(temp4)
-            return temp5
-
-    class Conv2dAvgPool2d(torch.nn.Module):
-        def __init__(
-            self,
-        ):
-            super().__init__()
-            self.conv = torch.nn.Conv2d(
-                3, 16, 7, bias=True, stride=2, padding=3, dilation=1
-            )
-            self.avgpool = torch.nn.AvgPool2d(3, stride=2, padding=1)
-
-        def forward(self, x):
-            temp1 = self.avgpool(self.conv(x))
-            return temp1
-
-    class Conv2dCatSameInputs(torch.nn.Module):
-        def __init__(
-            self,
-        ):
-            super().__init__()
-            self.conv = torch.nn.Conv2d(
-                3, 16, 7, bias=True, stride=2, padding=3, dilation=1
-            )
-            self.relu = torch.nn.ReLU()
-
-        def forward(self, x):
-            temp1 = self.relu(self.conv(x))
-            temp3 = torch.cat((temp1, temp1), 1)
-            return temp3
-
-    class Conv2dCatSingleInput(torch.nn.Module):
-        def __init__(
-            self,
-        ):
-            super().__init__()
-            self.conv = torch.nn.Conv2d(
-                3, 16, 7, bias=True, stride=2, padding=3, dilation=1
-            )
-            self.relu = torch.nn.ReLU()
-
-        def forward(self, x):
-            temp1 = self.relu(self.conv(x))
-            temp3 = torch.cat((temp1,), 1)
-            return temp3
 
     class SingleLinearModule(torch.nn.Module):
         def __init__(self, use_bias) -> None:
@@ -377,70 +180,6 @@ class TestHelperModules:
                     return tmp
                 else:
                     return self.linear(x) + self.linear2(x)
-
-    class LinearAddReLUModule(torch.nn.Module):
-        def __init__(
-            self,
-            inplace_add: bool = False,
-            linear_pos: NodePosType = NodePosType.left,
-            inplace_relu: bool = False,
-            use_bias: bool = False,
-        ) -> None:
-            super().__init__()
-            self.linear = torch.nn.Linear(
-                in_features=16, out_features=16, bias=use_bias
-            )
-            self.linear2 = torch.nn.Linear(
-                in_features=16, out_features=16, bias=use_bias
-            )
-            self.relu = nn.ReLU()
-            self.inplace_add = inplace_add
-            self.linear_pos = linear_pos
-            self.relu2 = nn.ReLU(inplace=inplace_relu)
-
-        def forward(self, x):
-            if self.linear_pos == NodePosType.left:
-                if self.inplace_add:
-                    tmp = self.linear(x)
-                    tmp += self.relu(x)
-                    return self.relu2(tmp)
-                else:
-                    tmp = self.linear(x)
-                    return self.relu2(tmp + self.relu(x))
-            elif self.linear_pos == NodePosType.right:
-                if self.inplace_add:
-                    tmp = self.relu(x)
-                    tmp += self.linear(x)
-                    return self.relu2(tmp)
-                else:
-                    return self.relu2(self.relu(x) + self.linear(x))
-            elif self.linear_pos == NodePosType.both:
-                if self.inplace_add:
-                    tmp = self.linear(x)
-                    tmp += self.linear2(x)
-                    return self.relu2(tmp)
-                else:
-                    return self.relu2(self.linear(x) + self.linear2(x))
-
-    class SerialsLinearAddReLUModule(torch.nn.Module):
-        """Serials of 2 Linear -> Add -> ReLU Pattern."""
-
-        def __init__(
-            self,
-        ) -> None:
-            super().__init__()
-            self.linear = torch.nn.Linear(in_features=16, out_features=16, bias=True)
-            self.linear2 = torch.nn.Linear(in_features=16, out_features=16, bias=True)
-            self.linear3 = torch.nn.Linear(in_features=16, out_features=16, bias=True)
-            self.linear4 = torch.nn.Linear(in_features=16, out_features=16, bias=True)
-            self.relu = nn.ReLU()
-            self.relu2 = nn.ReLU()
-
-        def forward(self, x):
-            x1 = self.linear(x)
-            res1 = self.relu(self.linear2(x1) + self.linear3(x1))
-            res2 = self.relu2(self.linear4(res1) + res1)
-            return res2
 
     class LinearAddModule2(torch.nn.Module):
         def __init__(
@@ -528,29 +267,6 @@ class TestHelperModules:
             weighted = torch.matmul(attention, v)
             return weighted
 
-    class Conv2dFlattenTranspose(nn.Module):
-        def __init__(self):
-            super().__init__()
-            self.projection = torch.nn.Conv2d(
-                3, 768, kernel_size=(16, 16), stride=(16, 16)
-            )
-            self.cls_token = torch.rand(1, 1, 768)
-
-        def forward(self, pixel_values):
-            embeddings = self.projection(pixel_values).flatten(2).transpose(1, 2)
-            embeddings = torch.cat((self.cls_token, embeddings), dim=1)
-            return embeddings
-
-    class Conv2dFlattenCatTranspose(nn.Module):
-        def __init__(self):
-            super().__init__()
-            self.conv = torch.nn.Conv2d(3, 768, kernel_size=(16, 16), stride=(16, 16))
-
-        def forward(self, x):
-            y = self.conv(x).flatten(2)
-            y = torch.cat([y, y], dim=-1)
-            return y.transpose(1, 2)
-
 
 class ArmInductorQuantTestCase(QuantizationTestCase):
     def _test_quantizer(
@@ -628,75 +344,6 @@ class TestQuantizePT2EArmInductor(ArmInductorQuantTestCase):
                 node_occurrence,
                 node_list,
             )
-
-    @skipIfNoArm
-    def cat(self):
-        """
-        Test pattern of conv2d with unary post ops (such as relu, hardtanh, hardswish, relu6) with ArmInductorQuantizer.
-        """
-        unary_map = {
-            "relu": [torch.nn.ReLU(inplace=False), torch.ops.aten.relu.default],
-            "relu_inplace": [torch.nn.ReLU(inplace=True), torch.ops.aten.relu_.default],
-            "hardtanh": [
-                torch.nn.Hardtanh(min_val=0.0, max_val=6.0, inplace=False),
-                torch.ops.aten.hardtanh.default,
-            ],
-            "hardtanh_inplace": [
-                torch.nn.Hardtanh(min_val=0.0, max_val=6.0, inplace=True),
-                torch.ops.aten.hardtanh_.default,
-            ],
-            "relu6": [torch.nn.ReLU6(inplace=False), torch.ops.aten.hardtanh.default],
-            "relu6_inplace": [
-                torch.nn.ReLU6(inplace=True),
-                torch.ops.aten.hardtanh_.default,
-            ],
-            "hardswish": [
-                torch.nn.Hardswish(inplace=False),
-                torch.ops.aten.hardswish.default,
-            ],
-            "hardswish_inplace": [
-                torch.nn.Hardswish(inplace=True),
-                torch.ops.aten.hardswish_.default,
-            ],
-            "swish": [torch.nn.SiLU(inplace=False), torch.ops.aten.silu.default],
-            "swish_inplace": [
-                torch.nn.SiLU(inplace=True),
-                torch.ops.aten.silu_.default,
-            ],
-        }
-        use_bias_list = [True, False]
-        with override_quantized_engine("arm"), torch.no_grad():
-            for unary_op, use_bias in itertools.product(
-                unary_map.keys(), use_bias_list
-            ):
-                m = TestHelperModules.Conv2dUnaryModule(
-                    unary_map[unary_op][0], use_bias=use_bias
-                ).eval()
-                example_inputs = (torch.randn(2, 3, 16, 16),)
-                quantizer = ArmInductorQuantizer().set_global(
-                    armiq.get_default_arm_inductor_quantization_config()
-                )
-                node_occurrence = {
-                    # one for input and weight of the conv
-                    torch.ops.quantized_decomposed.quantize_per_tensor.default: 3,
-                    torch.ops.quantized_decomposed.dequantize_per_tensor.default: 3,
-                    # note: quantize op for weights are const propagated
-                    torch.ops.quantized_decomposed.quantize_per_channel.default: 0,
-                    torch.ops.quantized_decomposed.dequantize_per_channel.default: 1,
-                }
-                node_list = [
-                    torch.ops.quantized_decomposed.quantize_per_tensor.default,
-                    torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-                    torch.ops.aten.conv2d.default,
-                    unary_map[unary_op][1],
-                ]
-                self._test_quantizer(
-                    m,
-                    example_inputs,
-                    quantizer,
-                    node_occurrence,
-                    node_list,
-                )
 
     @skipIfNoArm
     def test_conv2d_binary(self):
@@ -791,95 +438,6 @@ class TestQuantizePT2EArmInductor(ArmInductorQuantTestCase):
                     node_list,
                 )
 
-    @skipIfNoArm
-    def test_conv2d_binary_unary(self):
-        """
-        Test pattern of conv2d with binary + unary post ops (such as add + relu) with ArmInductorQuantizer.
-        Currently, only add as binary post op and relu as unary post op are supported.
-        """
-        conv2d_type_list = [NodePosType.left, NodePosType.both]
-        example_inputs = (torch.randn(2, 3, 6, 6),)
-        quantizer = ArmInductorQuantizer().set_global(
-            armiq.get_default_arm_inductor_quantization_config()
-        )
-        with override_quantized_engine("arm"), torch.no_grad():
-            for conv2d_type in conv2d_type_list:
-                m = TestHelperModules.Conv2dAddReLUModule(
-                    conv2d_type=conv2d_type,
-                ).eval()
-                if conv2d_type != NodePosType.both:
-                    node_occurrence = {
-                        # one for input for conv
-                        # one for extra input node of add
-                        torch.ops.quantized_decomposed.quantize_per_tensor.default: 2,
-                        torch.ops.quantized_decomposed.dequantize_per_tensor.default: 2,
-                        # note: quantize op for weights are const propagated
-                        torch.ops.quantized_decomposed.quantize_per_channel.default: 0,
-                        torch.ops.quantized_decomposed.dequantize_per_channel.default: 1,
-                    }
-                else:
-                    node_occurrence = {
-                        # one for input of the conv
-                        # one for input of another conv
-                        # 2 conv will share same input quant/dequant
-                        # one for extra input node of add
-                        torch.ops.quantized_decomposed.quantize_per_tensor.default: 2,
-                        torch.ops.quantized_decomposed.dequantize_per_tensor.default: 3,
-                        # note: quantize op for weights are const propagated
-                        torch.ops.quantized_decomposed.quantize_per_channel.default: 0,
-                        torch.ops.quantized_decomposed.dequantize_per_channel.default: 2,
-                    }
-                node_list = [
-                    torch.ops.quantized_decomposed.quantize_per_tensor.default,
-                    torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-                    torch.ops.aten.conv2d.default,
-                    torch.ops.aten.add.Tensor,
-                ]
-                self._test_quantizer(
-                    m,
-                    example_inputs,
-                    quantizer,
-                    node_occurrence,
-                    node_list,
-                )
-
-    @skipIfNoArm
-    def test_conv2d_serials_binary_unary(self):
-        """
-        Test pattern of 2 following up conv2d add relu with ArmInductorQuantizer.
-        """
-        with override_quantized_engine("arm"), torch.no_grad():
-            m = TestHelperModules.SerialsConv2dAddReLUModule().eval()
-            example_inputs = (torch.randn(2, 3, 16, 16),)
-            quantizer = ArmInductorQuantizer().set_global(
-                armiq.get_default_arm_inductor_quantization_config()
-            )
-            node_occurrence = {
-                torch.ops.quantized_decomposed.quantize_per_tensor.default: 4,
-                torch.ops.quantized_decomposed.dequantize_per_tensor.default: 6,
-                # quantize_per_channel for weights are const propagated
-                torch.ops.quantized_decomposed.quantize_per_channel.default: 0,
-                torch.ops.quantized_decomposed.dequantize_per_channel.default: 4,
-            }
-            node_list = [
-                torch.ops.quantized_decomposed.quantize_per_tensor.default,
-                torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-                torch.ops.aten.conv2d.default,
-                torch.ops.quantized_decomposed.quantize_per_tensor.default,
-                torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-                torch.ops.aten.conv2d.default,
-                torch.ops.aten.conv2d.default,
-                torch.ops.aten.add.Tensor,
-                torch.ops.aten.relu.default,
-            ]
-            self._test_quantizer(
-                m,
-                example_inputs,
-                quantizer,
-                node_occurrence,
-                node_list,
-            )
-
     def _single_op_share_observer_recipe_test_helper(self, m, x, single_op):
         quantizer = ArmInductorQuantizer().set_global(
             armiq.get_default_arm_inductor_quantization_config()
@@ -931,349 +489,6 @@ class TestQuantizePT2EArmInductor(ArmInductorQuantTestCase):
         self.assertTrue(isinstance(input_obs_of_conv, ObserverBase))
         self.assertTrue(input_obs_of_single_op is output_obs_of_single_op)
         self.assertTrue(input_obs_of_single_op is not input_obs_of_conv)
-
-    @skipIfNoArm
-    def test_maxpool2d_recipe(self):
-        r"""
-        Test pattern: int8_in_int8_out_ops(maxpool) - non_quantizable op(pow)
-        Since maxpool is a int8_in_int8_out_op, there is obs between maxpool and pow.
-        """
-        self._single_op_share_observer_recipe_test_helper(
-            TestHelperModules.Conv2dSingleOpPowModule(nn.MaxPool2d(1, 1)).eval(),
-            torch.rand(1, 2, 14, 14),
-            torch.ops.aten.max_pool2d.default,
-        )
-
-    @skipIfNoArm
-    def test_adaptive_avg_pool2d_recipe(self):
-        r"""
-        Test pattern: int8_in_int8_out_ops(adaptive_avg_pool2d) - non_quantizable op(pow)
-        Since adaptive_avg_pool2d is a int8_in_int8_out_op, there is obs between adaptive_avg_pool2d and pow.
-        """
-        self._single_op_share_observer_recipe_test_helper(
-            TestHelperModules.Conv2dSingleOpPowModule(
-                nn.AdaptiveAvgPool2d((1, 1))
-            ).eval(),
-            torch.rand(1, 2, 14, 14),
-            torch.ops.aten.adaptive_avg_pool2d.default,
-        )
-
-    @skipIfNoArm
-    def test_flatten_recipe(self):
-        r"""
-        Test pattern: conv -> flatten -> cat -> transpose
-        """
-        m = TestHelperModules.Conv2dFlattenCatTranspose().eval()
-        x = torch.randn(1, 3, 224, 224)
-        quantizer = ArmInductorQuantizer().set_global(
-            armiq.get_default_arm_inductor_quantization_config()
-        )
-        example_inputs = (x,)
-        node_occurrence = {
-            torch.ops.quantized_decomposed.quantize_per_tensor.default: 4,
-            torch.ops.quantized_decomposed.dequantize_per_tensor.default: 4,
-            # quantize_per_channel for weights are const propagated
-            torch.ops.quantized_decomposed.quantize_per_channel.default: 0,
-            torch.ops.quantized_decomposed.dequantize_per_channel.default: 1,
-        }
-        node_list = [
-            torch.ops.quantized_decomposed.quantize_per_tensor.default,
-            torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-            torch.ops.aten.conv2d.default,
-            torch.ops.quantized_decomposed.quantize_per_tensor.default,
-            torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-            torch.ops.aten.flatten.using_ints,
-            torch.ops.quantized_decomposed.quantize_per_tensor.default,
-            torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-            torch.ops.aten.cat.default,
-            torch.ops.quantized_decomposed.quantize_per_tensor.default,
-            torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-        ]
-        _, prepare_model, _ = self._test_quantizer(
-            m,
-            example_inputs,
-            quantizer,
-            node_occurrence,
-            node_list,
-        )
-        # Check Flatten has share observer at input and output
-        for node in prepare_model.graph.nodes:
-            if (
-                node.op == "call_function"
-                and node.target is torch.ops.aten.flatten.using_ints
-            ):
-                single_op_node = node
-                input_obs_of_single_op = getattr(
-                    prepare_model, single_op_node.args[0].target
-                )
-                output_obs_of_single_op = getattr(
-                    prepare_model, next(iter(single_op_node.users)).target
-                )
-            elif (
-                node.op == "call_function"
-                and node.target is torch.ops.aten.conv2d.default
-            ):
-                conv_node = node
-                input_obs_of_conv = getattr(prepare_model, conv_node.args[0].target)
-        self.assertTrue(isinstance(input_obs_of_single_op, ObserverBase))
-        self.assertTrue(isinstance(output_obs_of_single_op, ObserverBase))
-        self.assertTrue(isinstance(input_obs_of_conv, ObserverBase))
-        self.assertTrue(input_obs_of_single_op is output_obs_of_single_op)
-        self.assertTrue(input_obs_of_single_op is not input_obs_of_conv)
-
-    @skipIfNoArm
-    def test_flatten_recipe2(self):
-        r"""
-        Test pattern: conv -> flatten -> transpose
-        """
-        m = TestHelperModules.Conv2dFlattenTranspose().eval()
-        x = torch.randn(1, 3, 224, 224)
-        quantizer = ArmInductorQuantizer().set_global(
-            armiq.get_default_arm_inductor_quantization_config()
-        )
-        example_inputs = (x,)
-        node_occurrence = {
-            torch.ops.quantized_decomposed.quantize_per_tensor.default: 1,
-            torch.ops.quantized_decomposed.dequantize_per_tensor.default: 1,
-            # quantize_per_channel for weights are const propagated
-            torch.ops.quantized_decomposed.quantize_per_channel.default: 0,
-            torch.ops.quantized_decomposed.dequantize_per_channel.default: 1,
-        }
-        node_list = [
-            torch.ops.quantized_decomposed.quantize_per_tensor.default,
-            torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-            torch.ops.aten.conv2d.default,
-            torch.ops.aten.flatten.using_ints,
-            torch.ops.aten.transpose.int,
-        ]
-        self._test_quantizer(
-            m,
-            example_inputs,
-            quantizer,
-            node_occurrence,
-            node_list,
-        )
-
-    @skipIfNoArm
-    def test_cat_recipe(self):
-        r"""
-        Test pattern: conv -> cat -> maxpool2d
-        Since cat, maxpool is a int8_in_int8_out_op, the inputs and outputs should with same observer.
-        """
-        m = TestHelperModules.Conv2dCatMaxpool2d().eval()
-        x = torch.randn(16, 3, 16, 16).contiguous(memory_format=torch.channels_last)
-        quantizer = ArmInductorQuantizer().set_global(
-            armiq.get_default_arm_inductor_quantization_config()
-        )
-        example_inputs = (x,)
-        node_occurrence = {
-            torch.ops.quantized_decomposed.quantize_per_tensor.default: 6,
-            torch.ops.quantized_decomposed.dequantize_per_tensor.default: 6,
-            # quantize_per_channel for weights are const propagated
-            torch.ops.quantized_decomposed.quantize_per_channel.default: 0,
-            torch.ops.quantized_decomposed.dequantize_per_channel.default: 3,
-        }
-        node_list = [
-            torch.ops.quantized_decomposed.quantize_per_tensor.default,
-            torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-            torch.ops.aten.conv2d.default,
-            torch.ops.quantized_decomposed.quantize_per_tensor.default,
-            torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-            torch.ops.aten.cat.default,
-            torch.ops.quantized_decomposed.quantize_per_tensor.default,
-            torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-            torch.ops.aten.max_pool2d.default,
-            torch.ops.quantized_decomposed.quantize_per_tensor.default,
-            torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-        ]
-        _, prepare_model, _ = self._test_quantizer(
-            m,
-            example_inputs,
-            quantizer,
-            node_occurrence,
-            node_list,
-        )
-        # Check Cat/Maxpool2d has share observer at input and output
-        for node in prepare_model.graph.nodes:
-            if node.op == "call_function" and node.target == torch.ops.aten.cat.default:
-                cat_act_obs0 = getattr(prepare_model, node.all_input_nodes[0].target)
-                cat_act_obs1 = getattr(prepare_model, node.all_input_nodes[1].target)
-                cat_out_obs = getattr(prepare_model, next(iter(node.users)).target)
-            elif (
-                node.op == "call_function"
-                and node.target is torch.ops.aten.max_pool2d.default
-            ):
-                maxpool_node = node
-                input_obs_of_maxpool = getattr(
-                    prepare_model, maxpool_node.args[0].target
-                )
-                output_obs_of_maxpool = getattr(
-                    prepare_model, next(iter(maxpool_node.users)).target
-                )
-        self.assertTrue(isinstance(cat_act_obs0, ObserverBase))
-        self.assertTrue(isinstance(cat_act_obs1, ObserverBase))
-        self.assertTrue(isinstance(cat_out_obs, ObserverBase))
-        self.assertTrue(isinstance(input_obs_of_maxpool, ObserverBase))
-        self.assertTrue(isinstance(output_obs_of_maxpool, ObserverBase))
-        self.assertTrue(cat_act_obs0 is cat_act_obs1)
-        self.assertTrue(cat_act_obs0 is cat_out_obs)
-        self.assertTrue(cat_out_obs is input_obs_of_maxpool)
-        self.assertTrue(input_obs_of_maxpool is output_obs_of_maxpool)
-
-    @skipIfNoArm
-    def test_cat_recipe_same_inputs(self):
-        r"""
-        Test pattern: conv -> cat([input0, input0])
-        Since cat has 2 input node of same tensor, they should also be with same observer.
-        """
-        m = TestHelperModules.Conv2dCatSameInputs().eval()
-        x = torch.randn(16, 3, 16, 16).contiguous(memory_format=torch.channels_last)
-        quantizer = ArmInductorQuantizer().set_global(
-            armiq.get_default_arm_inductor_quantization_config()
-        )
-        example_inputs = (x,)
-        node_occurrence = {
-            torch.ops.quantized_decomposed.quantize_per_tensor.default: 3,
-            torch.ops.quantized_decomposed.dequantize_per_tensor.default: 3,
-            # quantize_per_channel for weights are const propagated
-            torch.ops.quantized_decomposed.quantize_per_channel.default: 0,
-            torch.ops.quantized_decomposed.dequantize_per_channel.default: 1,
-        }
-        node_list = [
-            torch.ops.quantized_decomposed.quantize_per_tensor.default,
-            torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-            torch.ops.aten.conv2d.default,
-            torch.ops.quantized_decomposed.quantize_per_tensor.default,
-            torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-            torch.ops.aten.cat.default,
-            torch.ops.quantized_decomposed.quantize_per_tensor.default,
-            torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-        ]
-        _, prepare_model, _ = self._test_quantizer(
-            m,
-            example_inputs,
-            quantizer,
-            node_occurrence,
-            node_list,
-        )
-        # Check Cat has share observer at input and output
-        for node in prepare_model.graph.nodes:
-            if node.op == "call_function" and node.target == torch.ops.aten.cat.default:
-                cat_act_obs0 = getattr(prepare_model, node.args[0][0].target)
-                cat_act_obs1 = getattr(prepare_model, node.args[0][1].target)
-                cat_out_obs = getattr(prepare_model, next(iter(node.users)).target)
-        self.assertTrue(isinstance(cat_act_obs0, ObserverBase))
-        self.assertTrue(isinstance(cat_act_obs1, ObserverBase))
-        self.assertTrue(isinstance(cat_out_obs, ObserverBase))
-        self.assertTrue(cat_act_obs0 is cat_act_obs1)
-        self.assertTrue(cat_act_obs0 is cat_out_obs)
-
-    @skipIfNoArm
-    def test_cat_recipe_single_input(self):
-        r"""
-        Test pattern: conv -> cat([input0,])
-        Since cat has 1 input node, they should also be with same observer.
-        """
-        m = TestHelperModules.Conv2dCatSingleInput().eval()
-        x = torch.randn(16, 3, 16, 16).contiguous(memory_format=torch.channels_last)
-        quantizer = ArmInductorQuantizer().set_global(
-            armiq.get_default_arm_inductor_quantization_config()
-        )
-        example_inputs = (x,)
-        node_occurrence = {
-            torch.ops.quantized_decomposed.quantize_per_tensor.default: 3,
-            torch.ops.quantized_decomposed.dequantize_per_tensor.default: 3,
-            # quantize_per_channel for weights are const propagated
-            torch.ops.quantized_decomposed.quantize_per_channel.default: 0,
-            torch.ops.quantized_decomposed.dequantize_per_channel.default: 1,
-        }
-        node_list = [
-            torch.ops.quantized_decomposed.quantize_per_tensor.default,
-            torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-            torch.ops.aten.conv2d.default,
-            torch.ops.quantized_decomposed.quantize_per_tensor.default,
-            torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-            torch.ops.aten.cat.default,
-            torch.ops.quantized_decomposed.quantize_per_tensor.default,
-            torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-        ]
-        _, prepare_model, _ = self._test_quantizer(
-            m,
-            example_inputs,
-            quantizer,
-            node_occurrence,
-            node_list,
-        )
-        # Check Cat has share observer at input and output
-        for node in prepare_model.graph.nodes:
-            if node.op == "call_function" and node.target == torch.ops.aten.cat.default:
-                cat_act_obs0 = getattr(prepare_model, node.args[0][0].target)
-                cat_out_obs = getattr(prepare_model, next(iter(node.users)).target)
-        self.assertTrue(isinstance(cat_act_obs0, ObserverBase))
-        self.assertTrue(isinstance(cat_out_obs, ObserverBase))
-        self.assertTrue(cat_act_obs0 is cat_out_obs)
-
-    @skipIfNoArm
-    def test_avg_pool2d_recipe(self):
-        r"""
-        Test pattern: conv -> AvgPool2d
-        Since AvgPool2d is a int8_in_int8_out_op, the inputs and outputs should with same observer.
-        """
-        m = TestHelperModules.Conv2dAvgPool2d().eval()
-        x = torch.randn(16, 3, 16, 16).contiguous(memory_format=torch.channels_last)
-        quantizer = ArmInductorQuantizer().set_global(
-            armiq.get_default_arm_inductor_quantization_config()
-        )
-        example_inputs = (x,)
-        node_occurrence = {
-            torch.ops.quantized_decomposed.quantize_per_tensor.default: 3,
-            torch.ops.quantized_decomposed.dequantize_per_tensor.default: 3,
-            # quantize_per_channel for weights are const propagated
-            torch.ops.quantized_decomposed.quantize_per_channel.default: 0,
-            torch.ops.quantized_decomposed.dequantize_per_channel.default: 1,
-        }
-        node_list = [
-            torch.ops.quantized_decomposed.quantize_per_tensor.default,
-            torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-            torch.ops.aten.conv2d.default,
-            torch.ops.quantized_decomposed.quantize_per_tensor.default,
-            torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-            torch.ops.aten.avg_pool2d.default,
-            torch.ops.quantized_decomposed.quantize_per_tensor.default,
-            torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-        ]
-        _, prepare_model, _ = self._test_quantizer(
-            m,
-            example_inputs,
-            quantizer,
-            node_occurrence,
-            node_list,
-        )
-        for node in prepare_model.graph.nodes:
-            if (
-                node.op == "call_function"
-                and node.target is torch.ops.aten.avg_pool2d.default
-            ):
-                avgpool_node = node
-                input_obs_of_avgpool = getattr(
-                    prepare_model, avgpool_node.args[0].target
-                )
-                output_obs_of_avgpool = getattr(
-                    prepare_model, next(iter(avgpool_node.users)).target
-                )
-            elif (
-                node.op == "call_function"
-                and node.target is torch.ops.aten.conv2d.default
-            ):
-                conv_node = node
-                output_obs_of_conv = getattr(
-                    prepare_model, next(iter(conv_node.users)).target
-                )
-        self.assertTrue(isinstance(input_obs_of_avgpool, ObserverBase))
-        self.assertTrue(isinstance(output_obs_of_avgpool, ObserverBase))
-        self.assertTrue(isinstance(output_obs_of_conv, ObserverBase))
-        self.assertTrue(input_obs_of_avgpool is output_obs_of_avgpool)
-        self.assertTrue(input_obs_of_avgpool is output_obs_of_conv)
 
     @skipIfNoArm
     def test_linear(self):
@@ -1539,306 +754,6 @@ class TestQuantizePT2EArmInductor(ArmInductorQuantTestCase):
                 }
                 self._check_annotation_stat(fq_m, expected_annotation_stat)
 
-    @skipIfNoArm
-    def test_linear_binary(self):
-        self._test_linear_binary_helper()
-
-    @skipIfNoArm
-    def test_linear_binary_qat(self):
-        self._test_linear_binary_helper(is_qat=True)
-
-    @skipIfNoArm
-    def test_linear_binary_dynamic(self):
-        self._test_linear_binary_helper(is_dynamic=True)
-
-    @skipIfNoArm
-    def test_linear_binary_dynamic_qat(self):
-        self._test_linear_binary_helper(is_qat=True, is_dynamic=True)
-
-    @skipIfNoArm
-    def test_linear_binary2(self):
-        """
-        Test Pattern:
-            tmp = linear_1(x)
-            tmp2 = linear_2(tmp)
-            return tmp + tmp2
-        Since linear_1 has 2 users, we should annotate linear_2 for binary fusion instead of linear_1
-        """
-        example_inputs = (torch.randn(2, 16),)
-        # TODO test for inplace add after refactoring of export_for_training
-        inplace_add_list = [False]
-        is_qat_list = [False, True]
-        is_dynamic_list = [False, True]
-        cases = itertools.product(inplace_add_list, is_qat_list, is_dynamic_list)
-        with override_quantized_engine("arm"), torch.no_grad():
-            for inplace_add, is_qat, is_dynamic in cases:
-                quantizer = ArmInductorQuantizer().set_global(
-                    armiq.get_default_arm_inductor_quantization_config(
-                        is_qat=is_qat, is_dynamic=is_dynamic
-                    )
-                )
-                m = TestHelperModules.LinearAddModule2(inplace_add=inplace_add).eval()
-                quantize_per_tensor_op = (
-                    torch.ops.quantized_decomposed.quantize_per_tensor.tensor
-                    if is_dynamic
-                    else torch.ops.quantized_decomposed.quantize_per_tensor.default
-                )
-                dequantize_per_tensor_op = (
-                    torch.ops.quantized_decomposed.dequantize_per_tensor.tensor
-                    if is_dynamic
-                    else torch.ops.quantized_decomposed.dequantize_per_tensor.default
-                )
-                # Two q-dq nodes for inputs of linear nodes
-                # No q-dq for extra input node of add
-                node_occurrence = {
-                    quantize_per_tensor_op: 2,
-                    dequantize_per_tensor_op: 2,
-                    # quantize_per_channel for weights are const propagated
-                    torch.ops.quantized_decomposed.quantize_per_channel.default: 0,
-                    torch.ops.quantized_decomposed.dequantize_per_channel.default: 2,
-                }
-                node_list = [
-                    torch.ops.quantized_decomposed.dequantize_per_channel.default,
-                    quantize_per_tensor_op,
-                    dequantize_per_tensor_op,
-                    torch.ops.aten.linear.default,
-                    (
-                        torch.ops.aten.add_.Tensor
-                        if inplace_add
-                        else torch.ops.aten.add.Tensor
-                    ),
-                ]
-                fq_m = self._test_quantizer(
-                    m,
-                    example_inputs,
-                    quantizer,
-                    node_occurrence,
-                    node_list,
-                )[-1]
-                # One linear and add are fused. The other linear is quantized alone if present
-                aten = torch.ops.aten
-                add_op = aten.add_.Tensor if inplace_add else aten.add.Tensor
-                expected_annotation_stat = {
-                    aten.linear.default: {
-                        "annotated": 2,
-                        "is_quant_out": 1,
-                    },
-                    add_op: {"annotated": 1, "is_quant_out": 1},
-                }
-                self._check_annotation_stat(fq_m, expected_annotation_stat)
-
-    @skipIfNoArm
-    def _test_linear_binary_unary_helper(self, is_qat=False, is_dynamic=False):
-        """
-        Test pattern of linear with binary + unary post ops (such as add + relu) with ArmInductorQuantizer.
-        Currently, only add as binary post op and relu as unary post op are supported.
-        """
-        linear_pos_list = [NodePosType.left, NodePosType.right, NodePosType.both]
-        # TODO test for inplace add after refactoring of export_for_training
-        inplace_add_list = [False]
-        # TODO test for inplace relu after refactoring of export_for_training
-        inplace_relu_list = [False]
-        example_inputs = (torch.randn(2, 16),)
-        quantizer = ArmInductorQuantizer().set_global(
-            armiq.get_default_arm_inductor_quantization_config(
-                is_qat=is_qat,
-                is_dynamic=is_dynamic,
-            )
-        )
-        quantize_per_tensor_op = (
-            torch.ops.quantized_decomposed.quantize_per_tensor.tensor
-            if is_dynamic
-            else torch.ops.quantized_decomposed.quantize_per_tensor.default
-        )
-        dequantize_per_tensor_op = (
-            torch.ops.quantized_decomposed.dequantize_per_tensor.tensor
-            if is_dynamic
-            else torch.ops.quantized_decomposed.dequantize_per_tensor.default
-        )
-        cases = itertools.product(linear_pos_list, inplace_add_list, inplace_relu_list)
-        with override_quantized_engine("arm"), torch.no_grad():
-            for linear_pos, inplace_add, inplace_relu in cases:
-                m = TestHelperModules.LinearAddReLUModule(
-                    inplace_add=inplace_add,
-                    linear_pos=linear_pos,
-                    inplace_relu=inplace_relu,
-                ).eval()
-                if linear_pos != NodePosType.both:
-                    node_occurrence = {
-                        # Only one q-dq node for input of the linear
-                        # No q-dq node for extra input node of add
-                        quantize_per_tensor_op: 1,
-                        dequantize_per_tensor_op: 1,
-                        # note: quantize op for weights are const propagated
-                        torch.ops.quantized_decomposed.quantize_per_channel.default: 0,
-                        torch.ops.quantized_decomposed.dequantize_per_channel.default: 1,
-                    }
-                else:
-                    # convert_pt2e disables duplicate dequant for dynamic quant
-                    num_dequant = 1 if is_dynamic else 2
-                    node_occurrence = {
-                        # One quantize_per_tensor for both linear nodes (shared)
-                        # Two dequantize_per_tensor for two linear nodes
-                        # No q-dq for extra input node of add
-                        quantize_per_tensor_op: 1,
-                        dequantize_per_tensor_op: num_dequant,
-                        # note: quantize op for weights are const propagated
-                        torch.ops.quantized_decomposed.quantize_per_channel.default: 0,
-                        torch.ops.quantized_decomposed.dequantize_per_channel.default: 2,
-                    }
-                node_list = [
-                    quantize_per_tensor_op,
-                    dequantize_per_tensor_op,
-                    torch.ops.aten.linear.default,
-                    (
-                        torch.ops.aten.add_.Tensor
-                        if inplace_add
-                        else torch.ops.aten.add.Tensor
-                    ),
-                ]
-                fq_m = self._test_quantizer(
-                    m,
-                    example_inputs,
-                    quantizer,
-                    node_occurrence,
-                    node_list,
-                )[-1]
-                # linear, add, relu are fused
-                # The other linear is quantized alone if present
-                aten = torch.ops.aten
-                add_op = aten.add_.Tensor if inplace_add else aten.add.Tensor
-                relu_op = aten.relu_.default if inplace_relu else aten.relu.default
-                expected_annotation_stat = {
-                    aten.linear.default: {
-                        "annotated": 2 if linear_pos == NodePosType.both else 1,
-                        "is_quant_out": 1 if linear_pos == NodePosType.both else 0,
-                    },
-                    add_op: {"annotated": 1, "is_quant_out": 0},
-                    relu_op: {"annotated": 1, "is_quant_out": 1},
-                }
-                self._check_annotation_stat(fq_m, expected_annotation_stat)
-
-    @skipIfNoArm
-    def test_linear_binary_unary(self):
-        self._test_linear_binary_unary_helper()
-
-    @skipIfNoArm
-    def test_linear_binary_unary_qat(self):
-        self._test_linear_binary_unary_helper(is_qat=True)
-
-    @skipIfNoArm
-    def test_linear_binary_unary_dynamic(self):
-        self._test_linear_binary_unary_helper(is_dynamic=True)
-
-    @skipIfNoArm
-    def test_linear_binary_unary_dynamic_qat(self):
-        self._test_linear_binary_unary_helper(is_qat=True, is_dynamic=True)
-
-    @skipIfNoArm
-    def test_linear_binary_unary_serials(self):
-        """
-        Test pattern of 2 following up linear add relu with ArmInductorQuantizer.
-        """
-        is_qat_list = [False, True]
-        is_dynamic_list = [False, True]
-        cases = itertools.product(is_qat_list, is_dynamic_list)
-        with override_quantized_engine("arm"), torch.no_grad():
-            for is_qat, is_dynamic in cases:
-                m = TestHelperModules.SerialsLinearAddReLUModule().eval()
-                example_inputs = (torch.randn(2, 16),)
-                quantizer = ArmInductorQuantizer().set_global(
-                    armiq.get_default_arm_inductor_quantization_config(
-                        is_qat=is_qat,
-                        is_dynamic=is_dynamic,
-                    )
-                )
-                quantize_per_tensor_op = (
-                    torch.ops.quantized_decomposed.quantize_per_tensor.tensor
-                    if is_dynamic
-                    else torch.ops.quantized_decomposed.quantize_per_tensor.default
-                )
-                dequantize_per_tensor_op = (
-                    torch.ops.quantized_decomposed.dequantize_per_tensor.tensor
-                    if is_dynamic
-                    else torch.ops.quantized_decomposed.dequantize_per_tensor.default
-                )
-                # convert_pt2e disables duplicate dequant for dynamic quant
-                num_dequant = 3 if is_dynamic else 4
-                node_occurrence = {
-                    # quantize_per_tensor: 1 for linear_1, 1 for linear_2/3 (shared), 1 for linear_4
-                    # dequantize_per_tensor: 1 for each linear
-                    # No q-dq for extra input node of add
-                    quantize_per_tensor_op: 3,
-                    dequantize_per_tensor_op: num_dequant,
-                    # quantize_per_channel for weights are const propagated
-                    torch.ops.quantized_decomposed.quantize_per_channel.default: 0,
-                    torch.ops.quantized_decomposed.dequantize_per_channel.default: 4,
-                }
-                node_list = [
-                    torch.ops.quantized_decomposed.dequantize_per_channel.default,
-                    quantize_per_tensor_op,
-                    dequantize_per_tensor_op,
-                    torch.ops.aten.linear.default,
-                    torch.ops.aten.linear.default,
-                    torch.ops.aten.linear.default,
-                    torch.ops.aten.add.Tensor,
-                    torch.ops.aten.relu.default,
-                ]
-                fq_m = self._test_quantizer(
-                    m,
-                    example_inputs,
-                    quantizer,
-                    node_occurrence,
-                    node_list,
-                )[-1]
-                # Two linear nodes are quantized alone
-                # The other two are fused with add and relu
-                aten = torch.ops.aten
-                expected_annotation_stat = {
-                    aten.linear.default: {
-                        "annotated": 4,
-                        "is_quant_out": 2,
-                    },
-                    aten.add.Tensor: {"annotated": 2, "is_quant_out": 0},
-                    aten.relu.default: {"annotated": 2, "is_quant_out": 2},
-                }
-                self._check_annotation_stat(fq_m, expected_annotation_stat)
-
-    @skipIfNoArm
-    def test_linear_dynamic_fp16(self):
-        """
-        Test pattern of linear_dynamic_fp16.
-        """
-        with override_quantized_engine("arm"), torch.no_grad():
-            for use_bias in [True, False]:
-                m = TestHelperModules.SingleLinearModule(use_bias).eval()
-                example_inputs = (torch.randn(2, 4),)
-                quantizer = ArmInductorQuantizer().set_global(
-                    armiq.get_default_arm_inductor_quantization_config()
-                )
-                quantizer.set_module_type_qconfig(
-                    torch.nn.Linear, armiq.get_arm_inductor_linear_dynamic_fp16_config()
-                )
-                node_occurrence = {
-                    # 2 convert_element_type nodes are inserted for weight
-                    torch.ops.quantized_decomposed.quantize_per_tensor.default: 0,
-                    torch.ops.quantized_decomposed.dequantize_per_tensor.default: 0,
-                    torch.ops.quantized_decomposed.quantize_per_channel.default: 0,
-                    torch.ops.quantized_decomposed.dequantize_per_channel.default: 0,
-                    torch.ops.quantized_decomposed.convert_element_type.no_fuse: 2,
-                }
-                node_list = [
-                    torch.ops.quantized_decomposed.convert_element_type.no_fuse,
-                    torch.ops.aten.linear.default,
-                ]
-                self._test_quantizer(
-                    m,
-                    example_inputs,
-                    quantizer,
-                    node_occurrence,
-                    node_list,
-                )
-
     @skipIfTorchDynamo("very slow")
     @skipIfNoArm
     def test_qat_conv2d(self):
@@ -1876,80 +791,6 @@ class TestQuantizePT2EArmInductor(ArmInductorQuantTestCase):
                 node_list,
                 is_qat=True,
             )
-
-    @skipIfTorchDynamo("very slow")
-    @skipIfNoArm
-    def test_qat_conv2d_unary(self):
-        """
-        Test QAT pattern of conv2d_bn with unary post ops (such as relu, sigmoid) with ArmInductorQuantizer.
-        Currently, only relu as unary post op is supported.
-        """
-        unary_map = {
-            "relu": [torch.nn.ReLU(inplace=False), torch.ops.aten.relu.default],
-            "relu_inplace": [torch.nn.ReLU(inplace=True), torch.ops.aten.relu_.default],
-            "hardtanh": [
-                torch.nn.Hardtanh(min_val=0.0, max_val=6.0, inplace=False),
-                torch.ops.aten.hardtanh.default,
-            ],
-            "hardtanh_inplace": [
-                torch.nn.Hardtanh(min_val=0.0, max_val=6.0, inplace=True),
-                torch.ops.aten.hardtanh_.default,
-            ],
-            "relu6": [torch.nn.ReLU6(inplace=False), torch.ops.aten.hardtanh.default],
-            "relu6_inplace": [
-                torch.nn.ReLU6(inplace=True),
-                torch.ops.aten.hardtanh_.default,
-            ],
-            "hardswish": [
-                torch.nn.Hardswish(inplace=False),
-                torch.ops.aten.hardswish.default,
-            ],
-            "hardswish_inplace": [
-                torch.nn.Hardswish(inplace=True),
-                torch.ops.aten.hardswish_.default,
-            ],
-            "swish": [torch.nn.SiLU(inplace=False), torch.ops.aten.silu.default],
-            "swish_inplace": [
-                torch.nn.SiLU(inplace=True),
-                torch.ops.aten.silu_.default,
-            ],
-        }
-
-        with override_quantized_engine("arm"):
-            for unary_op in unary_map.keys():
-                m = TestHelperModules.Conv2dUnaryModule(
-                    unary_map[unary_op][0], with_bn=True
-                )
-                example_inputs = (torch.randn(2, 3, 16, 16),)
-                quantizer = ArmInductorQuantizer().set_global(
-                    armiq.get_default_arm_inductor_quantization_config(is_qat=True)
-                )
-                node_occurrence = {
-                    # one for input and weight of the conv, one for output for the relu
-                    torch.ops.quantized_decomposed.quantize_per_tensor.default: 3,
-                    torch.ops.quantized_decomposed.dequantize_per_tensor.default: 3,
-                    # note: quantize op for weights are const propagated
-                    torch.ops.quantized_decomposed.quantize_per_channel.default: 0,
-                    torch.ops.quantized_decomposed.dequantize_per_channel.default: 1,
-                    # BN should be folded into Conv
-                    torch.ops.aten._native_batch_norm_legit.default: 0,
-                }
-                node_list = [
-                    torch.ops.quantized_decomposed.quantize_per_tensor.default,
-                    torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-                    torch.ops.aten.conv2d.default,
-                    unary_map[unary_op][1],
-                    torch.ops.quantized_decomposed.quantize_per_tensor.default,
-                    torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-                ]
-                self._test_quantizer(
-                    m,
-                    example_inputs,
-                    quantizer,
-                    node_occurrence,
-                    node_list,
-                    is_qat=True,
-                )
 
     @skipIfTorchDynamo("very slow")
     @skipIfNoArm
@@ -2046,48 +887,6 @@ class TestQuantizePT2EArmInductor(ArmInductorQuantTestCase):
                     node_list,
                     is_qat=True,
                 )
-
-    @skipIfTorchDynamo("very slow")
-    @skipIfNoArm
-    def test_qat_conv2d_binary_unary(self):
-        """
-        Test QAT pattern of conv2d_bn with binary + unary post ops (such as add + relu) with ArmInductorQuantizer.
-        Currently, only add as binary post op and relu as unary post op are supported.
-        """
-        example_inputs = (torch.randn(2, 3, 6, 6),)
-        quantizer = ArmInductorQuantizer().set_global(
-            armiq.get_default_arm_inductor_quantization_config(is_qat=True)
-        )
-        with override_quantized_engine("arm"):
-            m = TestHelperModules.Conv2dAddReLUModule(with_bn=True)
-            node_occurrence = {
-                # one for input for conv
-                # one for output for the relu
-                # one for extra input node of add
-                torch.ops.quantized_decomposed.quantize_per_tensor.default: 3,
-                torch.ops.quantized_decomposed.dequantize_per_tensor.default: 3,
-                # note: quantize op for weights are const propagated
-                torch.ops.quantized_decomposed.quantize_per_channel.default: 0,
-                torch.ops.quantized_decomposed.dequantize_per_channel.default: 1,
-                # BN should be folded into Conv
-                torch.ops.aten._native_batch_norm_legit.default: 0,
-            }
-            node_list = [
-                torch.ops.quantized_decomposed.quantize_per_tensor.default,
-                torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-                torch.ops.aten.conv2d.default,
-                torch.ops.aten.add.Tensor,
-                torch.ops.quantized_decomposed.quantize_per_tensor.default,
-                torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-            ]
-            self._test_quantizer(
-                m,
-                example_inputs,
-                quantizer,
-                node_occurrence,
-                node_list,
-                is_qat=True,
-            )
 
     @skipIfNoArm
     def test_dynamic_quant_linear(self):
@@ -2620,41 +1419,6 @@ class TestQuantizePT2EArmInductor(ArmInductorQuantTestCase):
             node_list = [
                 torch.ops.aten.linear.default,
                 torch.ops.aten.relu.default,
-            ]
-            self._test_quantizer(
-                m,
-                example_inputs,
-                quantizer,
-                node_occurrence,
-                node_list,
-            )
-
-    @skipIfNoArm
-    def test_filter_maxpool2d_recipe(self):
-        """
-        Test removing maxpool2d from default recipe of ArmInductorQuantizer.
-        """
-        with override_quantized_engine("arm"), torch.no_grad():
-            m = TestHelperModules.Conv2dUnaryModule(torch.nn.ReLU(inplace=False)).eval()
-            example_inputs = (torch.randn(2, 3, 16, 16),)
-            quantizer = ArmInductorQuantizer().set_global(
-                armiq.get_default_arm_inductor_quantization_config()
-            )
-            quantizer.set_function_type_qconfig(torch.nn.functional.max_pool2d, None)
-            node_occurrence = {
-                # one for input and weight of the conv
-                torch.ops.quantized_decomposed.quantize_per_tensor.default: 1,
-                torch.ops.quantized_decomposed.dequantize_per_tensor.default: 1,
-                # note: quantize op for weights are const propagated
-                torch.ops.quantized_decomposed.quantize_per_channel.default: 0,
-                torch.ops.quantized_decomposed.dequantize_per_channel.default: 1,
-            }
-            node_list = [
-                torch.ops.quantized_decomposed.quantize_per_tensor.default,
-                torch.ops.quantized_decomposed.dequantize_per_tensor.default,
-                torch.ops.aten.conv2d.default,
-                torch.ops.aten.relu.default,
-                torch.ops.aten.max_pool2d.default,
             ]
             self._test_quantizer(
                 m,
